tests/wpt/tests/css/css-fonts/font-variant-alternates-parsing.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test:  font-variant-alternates: historical-forms; parses case-insensitively</title>
<link rel="author" title="Emilio Cobos Ãlvarez" href="emilio@crisal.io">
<link rel="help" href="https://www.w3.org/TR/css-fonts-4/#font-variant-alternates-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div></div>
<script>
test(function() {
  let div = document.querySelector('div');
  div.style.fontVariantAlternates = "Historical-Forms";
  assert_equals(
    getComputedStyle(div).fontVariantAlternates,
    "historical-forms",
    "historical-forms is parsed case-insensitively"
  );
});
</script>
